requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
scikit-learn>=1.6.0
numpy>=1.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pymorphy2>=0.9.1
setuptools>=79.0.0
backports.inspect>=0.0.3 